requirements.txt
aiohappyeyeballs == 2.6.1
aiohttp == 3.12.13
aiosignal == 1.3.2
aiosqlite == 0.21.0
attrs == 25.3.0
Base_69 == 0.2.2
ffmpeg-python == 0.2.0
frozenlist == 1.7.0
future == 1.0.0
idna == 3.10
imageio-ffmpeg == 0.6.0
multidict == 6.6.3
numpy == 2.3.1
pip == 23.2.1
propcache == 0.3.2
py-cord == 2.6.1
simpcalci == 0.0.2
typing_extensions == 4.14.0
urllib3 == 2.5.0
yarl == 1.20.1
python-dotenv
regex 
feedparser
dateparser
httpx
pillow